requirements.txt
# ============================================
# AI Study Companion - Requirements
# ============================================

# --- Core ML Framework ---
torch>=2.1.0
transformers>=4.36.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# --- HuggingFace Inference API ---
huggingface-hub>=0.20.0

# --- PDF Processing ---
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# --- Image OCR ---
pytesseract>=0.3.10
Pillow>=10.0.0

# --- Web Framework ---
flask>=3.0.0
flask-cors>=4.0.0

# --- NLP Utilities ---
nltk>=3.8.0
langdetect>=1.0.9

# --- Evaluation Metrics ---
rouge-score>=0.1.2
scikit-learn>=1.3.0

# --- Keyword Extraction (fallback) ---
rake-nltk>=1.0.6
yake>=0.4.8

# --- General ---
numpy>=1.24.0
fpdf2>=2.7.0